🐛 [fix] : spinner 안되는 오류 해결
